Rename `fixtures_server.py` into `fixtures_server_v2.py`
