+++ tests/v1/01_unit/test_unit_config.py
@@ -7,7 +7,7 @@ from fractal_server.config import FractalConfigurationError
 from fractal_server.config import OAuthClientConfig
 from fractal_server.config import Settings
 from fractal_server.syringe import Inject
-from tests.fixtures_server import DB_ENGINE
+from tests.fixtures_server_v1 import DB_ENGINE
 
 
 def test_settings_injection(override_settings):
